requirements.txt
accelerate==0.30.0
annotated-types==0.6.0
anyio==4.3.0
certifi==2024.2.2
charset-normalizer==3.3.2
decord==0.6.0
distro==1.9.0
einops==0.8.0
et-xmlfile==1.1.0
exceptiongroup==1.2.1
filelock==3.14.0
#flash-attn==2.6.3
fsspec==2024.3.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.24.6
idna==3.7
Jinja2==3.1.4
joblib==1.4.2
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
#nvidia-cuda-runtime-cu12==12.1.105
#nvidia-cudnn-cu12==8.9.2.26
#nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
#nvidia-cusolver-cu12==11.4.5.107
#nvidia-cusparse-cu12==12.1.0.106
#nvidia-nccl-cu12==2.20.5
#nvidia-nvjitlink-cu12==12.4.127
#nvidia-nvtx-cu12==12.1.105
openai==1.26.0
openpyxl==3.1.3
packaging==24.0
pandas==2.2.2
peft==0.13.0
pillow==10.3.0
protobuf==5.28.0
psutil==5.9.8
pydantic==2.7.1
pydantic_core==2.18.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
qwen-vl-utils==0.0.4
regex==2024.4.28
requests==2.31.0
safetensors==0.4.3
scikit-learn==1.5.2
scipy==1.13.1
sentencepiece==0.2.0
six==1.16.0
sniffio==1.3.1
sympy==1.12
threadpoolctl==3.5.0
tiktoken==0.6.0
timm==1.0.9
tokenizers==0.19.1
torch==2.3.0
torchaudio==2.3.0
torchvision==0.18.0
tqdm==4.66.4
transformers @ git+https://github.com/huggingface/transformers@f24f08432960023129817328cac6c01151d1f70f
#triton==2.3.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
